fixup! Rewrite trust USB restriction handling
For our Trust preferences, if the `trust_restrict_usb` setting is
missing, we default to assuming a value of 0 (don't restrict USB
devices). However, here in fw/b, we're instead defaulting to assuming a
value of 1 (restruct USB devices when device is locked). As such, this
can cause confusion because the user is expecting that we aren't
restricting USB connections by default.

Let's update the fw/b side of things to also assume a value of 0 instead
to clear up the confusion.

Change-Id: Ifa45bfb05ad4e745266f3da11ee3f0f4a6a3384b